% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_paths.R
\name{get_course_paths}
\alias{get_course_paths}
\title{Get Course Paths}
\usage{
get_course_paths(api_token)
}
\arguments{
\item{api_token}{Your personalized token provided by 'Skilljar'}
}
\value{
A data frame with users and user data
}
\description{
Returns a data.frame with all paths and the items in those paths.
}
\details{
Utilizing the API requires a token. This must be obtained by logging in
at dashboard.skilljar.com and going to Organization -> API Credentials.
There are different strategies for storing api tokens securely. It is 
an unnecessary risk to store the token in the script!
}
\examples{
\dontrun{
# Get course progress data
paths <- get_course_paths(api_token = "my-token")
}
}
\seealso{
See \url{https://api.skilljar.com/docs/} for documentation on
the 'Skilljar' API.
}
